ajuste backend e retirado front

+++ financeiro/src/main/java/com/rosetim/financeiro/service/cartao/CartaoService.java
@@ -1,13 +1,17 @@
 package com.rosetim.financeiro.service.cartao;
 
 import com.rosetim.financeiro.entity.cartao.CartaoEntity;
+import com.rosetim.financeiro.entity.gastos.GastosEntity;
+import com.rosetim.financeiro.exception.CustomException;
 import com.rosetim.financeiro.repository.cartao.CartaoRepository;
+import com.rosetim.financeiro.repository.gastos.GastosRepository;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -17,6 +21,9 @@ public class CartaoService {
 
     @Autowired
     CartaoRepository cartaoRepository;
+
+    @Autowired
+    GastosRepository gastosRepository;
 
     public CartaoEntity save(CartaoEntity cartaoEntity) throws Exception {
         try {
@@ -44,7 +51,16 @@ public class CartaoService {
 
     public void deleteById(Long cdCartao)throws Exception{
         try{
-            cartaoRepository.deleteById(cdCartao);
+            CartaoEntity cartao = cartaoRepository.findByCdCartao(cdCartao);
+
+            List<GastosEntity> gastosVinculados = gastosRepository.findByCartao(cartao);
+
+            if (gastosVinculados.size() <= 0){
+                cartaoRepository.deleteById(cdCartao);
+            }else {
+                throw new CustomException("Cartão vinculado a um gasto, não pode ser excluido");
+            }
+
         }catch (Exception e){
             throw new Exception(e);
         }
